feat: static binaries for Pi Zero/1-5, Omega 1/2+ in CI
Fix int32/int64 mismatch in Rusage CPU sampling for 32-bit archs
(MIPS, ARMv6/v7). Add CI matrix entries for:
- linux/armv6 (Pi Zero, Pi 1)
- linux/armv7 (Pi 2/3/4 32-bit)
- linux/mipsle softfloat (Onion Omega2+)
- linux/mips softfloat (Onion Omega 1)

All builds are CGO_ENABLED=0 fully static binaries (~12-14MB).

+++ cloud/internal/dashboard/state.go
@@ -70,8 +70,8 @@ func NewState() *State {
 	// Initialize CPU baseline.
 	var ru syscall.Rusage
 	if syscall.Getrusage(syscall.RUSAGE_SELF, &ru) == nil {
-		s.prevCPUUser = ru.Utime.Sec*1e6 + int64(ru.Utime.Usec)
-		s.prevCPUSys = ru.Stime.Sec*1e6 + int64(ru.Stime.Usec)
+		s.prevCPUUser = int64(ru.Utime.Sec)*1e6 + int64(ru.Utime.Usec)
+		s.prevCPUSys = int64(ru.Stime.Sec)*1e6 + int64(ru.Stime.Usec)
 	}
 	return s
 }
@@ -358,8 +358,8 @@ func (s *State) ToJSON() []byte {
 	var cpuPcts []int
 	var ru syscall.Rusage
 	if syscall.Getrusage(syscall.RUSAGE_SELF, &ru) == nil {
-		userUs := ru.Utime.Sec*1e6 + int64(ru.Utime.Usec)
-		sysUs := ru.Stime.Sec*1e6 + int64(ru.Stime.Usec)
+		userUs := int64(ru.Utime.Sec)*1e6 + int64(ru.Utime.Usec)
+		sysUs := int64(ru.Stime.Sec)*1e6 + int64(ru.Stime.Usec)
 		elapsed := time.Since(s.prevCPUTime).Microseconds()
 		if elapsed > 0 {
 			cpuUs := (userUs - s.prevCPUUser) + (sysUs - s.prevCPUSys)
